requirements.txt
Babel==2.12.1
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
Pillow==10.0.0
python-decouple==3.8
python-dotenv==1.0.0
twilio==8.7.0
rembg==2.0.53
psycopg==3.1.14
tensorflow==2.13.0
tensorflow_hub==0.15.0